SERVER-67852 Partial statistics pipeline

+++ jstests/noPassthrough/cluster_analyze_command.js
@@ -1,6 +1,8 @@
 (function() {
 "use strict";
 
+load("jstests/libs/sbe_util.js");  // For checkSBEEnabled.
+
 const st = new ShardingTest({
     shards: 2,
     mongos: 1,
@@ -11,6 +13,13 @@ const st = new ShardingTest({
 });
 
 const db = st.getDB("test");
+
+if (checkSBEEnabled(db, ["featureFlagSbeFull"], true)) {
+    jsTestLog("Skipping the test because it doesn't work in Full SBE...");
+    st.stop();
+    return;
+}
+
 const coll = db.analyze_coll;
 coll.drop();
 
@@ -22,13 +31,13 @@ assert.commandWorked(
 assert.commandWorked(coll.insert({a: 1, b: 2}));
 
 let res = db.runCommand({analyze: coll.getName()});
-assert.commandFailedWithCode(res, ErrorCodes.NotImplemented);
+assert.commandWorked(res);
 
 res = db.runCommand({analyze: coll.getName(), apiVersion: "1", apiStrict: true});
 assert.commandFailedWithCode(res, ErrorCodes.APIStrictError);
 
 res = db.runCommand({analyze: coll.getName(), writeConcern: {w: 1}});
-assert.commandFailedWithCode(res, ErrorCodes.NotImplemented);
+assert.commandWorked(res);
 
 st.stop();
 })();
